refactor: promote turn-queue/state/turn-store from telegram/ to shared channels/ modules
The lark channel (next commits) is the second stateful chat channel; the
three mechanisms it shares with telegram move up a level:

- channels/turn-queue.ts: already channel-neutral (label option) — moved verbatim.
- channels/state.ts: atomic state files — labels neutralized (the path names the channel).
- channels/turn-store.ts: the L1 durable-intent semantics, now generic over the
  record shape ({ label, isRecord, order }); telegram/turn-store.ts keeps the
  telegram-shaped record + validator + update_id arrival order over it, so its
  import surface (and the on-disk format) is unchanged.
