+++ src/channels/telegram/telegram.ts
@@ -46,9 +46,9 @@ import {
   telegramEnvelope,
 } from "./parse.ts";
 import { type TelegramFailure, defaultErrorMessage, streamReply } from "./preview.ts";
-import { ensureStateHome } from "./state.ts";
+import { ensureStateHome } from "../state.ts";
 import { type Target, callApi, editMessageText, sendMessage } from "./telegram-api.ts";
-import { createTurnQueue } from "./turn-queue.ts";
+import { createTurnQueue } from "../turn-queue.ts";
 import { type StoredTurn, createTurnStore } from "./turn-store.ts";
 
 // Re-export the public surface authored elsewhere, so `@fastagent-sh/fastagent/telegram` keeps one entry point.
